Added Practice Session functionality
As I'm awaiting data to train the model, the two initially planned sprints have been reversed, so the addition of user practice sessions marks the end of the initially proposed second sprint. In this sprint, I added the Practice Session feature (which is still technically incomplete without the actual pronunciation scoring model) - more specifically I added in the Javascript to detect breaks in speech, record audio and send it back to the backend. I also added a view in my backend to allow users to hear the correct pronunciation of their question and words in their response (by hovering over them).