Added authentication validation.
This feature is still part of the first print, and replaces the simplicity of the data sanitation present in the login and signup views with validation functions to ensure that all fields match the criteria mentioned in the data dictionary. Length validation is also implemented in the HTML <input> fields to limit users responses to the required length.