Trained model and attempted integration into Practice Session functionality.
This marks the end of the second (formerly the first) sprint of this project. In this sprint, once data was received, it was processed through an appropriate pipeline before being included in the  Siamese Neural Network training set. Through rigorous training and alteration, the model's loss was brought down to roughly 0.14 but due to delays in receiving data and the project complexity slightly exceeding pre-defined constraints, there was not sufficient time to integrate this model into the remaining application and resolve any errors that occurred when attempting to do so.